Portfolio Construction App v1.0
A locally hosted application that allows investors to construct a portfolio based on a set of pre-defined criteria that fit their investment strategy. Investors are allowed to choose the number of funds within a portfolio, the number of funds managed by emerging and established managers, their desired portfolio return and their risk appetite.